Added minors
